Minor reorganization of stray unit tests
